requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3

# Database
psycopg2-binary==2.9.7

# Image processing
Pillow==10.0.1

# Environment variables
python-decouple==3.8

# Date utilities - CRITICAL FOR LOANS
python-dateutil==2.8.2

# Development tools (minimal)
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0